Conflicts:
	org.zenframework.z8.server/src/main/java/org/zenframework/z8/server/ie/Export.java

+++ org.zenframework.z8.server/src/main/java/org/zenframework/z8/server/utils/NumericUtils.java
@@ -1,8 +1,8 @@
 package org.zenframework.z8.server.utils;
 
 public class NumericUtils {
-	static final public long Kilobyte = 1024;
-	static final public long Megabyte = Kilobyte * Kilobyte;
+	static final public int Kilobyte = 1024;
+	static final public int Megabyte = Kilobyte * Kilobyte;
 	static final public long Gigabyte = Megabyte * Kilobyte;
 
 }
